Delete old tracking events types
