+++ packages/ui/src/components/tracking/index.ts
@@ -1,5 +1,5 @@
 export * from './Consent';
+export * from './posthog-events';
 export * from './posthog-pageview';
 export * from './posthog-provider';
-export * from './trackingPlan';
 export * from './use-event-tracker';
