+++ packages/ui/src/components/tracking/posthog-events.ts
@@ -1,0 +1,120 @@
+export const Event = {
+  auth: {
+    submitRejoindreCollectivite: 'cta_submit',
+    submitForgottenPassword: 'cta_submit',
+    submitLogin: 'cta_submit',
+    submitResetPassword: 'cta_submit',
+    submitResendMessage: 'cta_submit',
+    submitSignup: 'cta_submit',
+    submitSignupDcp: 'cta_submit',
+    submitVerifyOTP: 'cta_submit',
+    viewAvecMdp: 'onglet_avec_mdp',
+    viewSansMdp: 'onglet_sans_mdp',
+  },
+  updateFiltres: 'filtres',
+  saveScore: 'referentiel_score_save',
+  paginationClick: 'pagination_click',
+  panier: {
+    createPlanClick: 'cta_creer_le_plan_click',
+    ctaPanierClick: 'cta_panier_click',
+    filtre: 'panier:filtre',
+    copierPanierUrl: 'copier_panier_URL',
+    clickAjout: 'ajout',
+    clickRetrait: 'retrait',
+    changeStatut: 'statut',
+    validerCreationPanierClick: 'cta_valider_creation_panier_click',
+    selectTab: 'onglet_panier',
+  },
+  indicateurs: {
+    createIndicateurPerso: 'cta_indicateur_perso_fa',
+    downloadXlsx: 'export_xlsx_telechargement',
+    deleteIndicateur: 'indicateur_suppression',
+    viewIndicateursList: 'explorerIndicateursClick',
+  },
+  plans: {
+    sideNavAjouterPlanClick: 'plansAction:side-nav-ajouter-plan-click',
+    createPlan: 'cta_creer',
+    importPlan: 'cta_importer',
+    startPanier: 'cta_commencer_pai',
+    exportPlan: 'export_plan',
+  },
+  fiches: {
+    updatePilotesGroupe: 'editer_personne_pilote_groupe',
+    updatePlanningGroupe: 'associer_planning_groupe',
+    updatePrioriteGroupe: 'associer_priorite_groupe',
+    updateStatutGroupe: 'associer_statut_groupe',
+    updateTagsLibresGroupe: 'associer_tags_perso_groupe',
+    exportPdf: 'export_PDF',
+    exportPdfGroupe: 'export_PDF_telechargement_groupe',
+    updateActeurs: 'validation_modale_acteurs_fa',
+    updateDescription: 'validation_modale_modifier_fa',
+    updatePlanning: 'validation_modale_planning_fa',
+    viewImpactInfo: 'cta_fa_fai',
+    downloadModele: 'cta_telecharger_modele',
+  },
+  trajectoire: {
+    selectIndicateur: 'indicateurs_trajectoire_snbc_select',
+    selectSecteur: 'selection_secteur',
+    downloadSnbcFile: 'indicateurs_trajectoire_snbc_file_download',
+    triggerSnbcCalculation: 'indicateurs_trajectoire_snbc_calculation_trigger',
+  },
+  tdb: {
+    updateFiltresIndicateurs: 'tdb_modifier_filtres_indicateurs',
+    updateFiltresActionsPilotes: 'tdb_modifier_filtres_actions_pilotes',
+    updateFiltresActionsModifiees: 'tdb_modifier_filtres_actions_modifiees',
+    updateFiltresCountByActions: 'tdb_modifier_filtres_count_by_actions',
+    updateFiltresSuiviPlans: 'tdb_modifier_filtres_suivi_plans',
+    validateFiltresIndicateurs: 'tdb_valider_filtres_indicateurs',
+    validateModulePerso: 'tdb_valider_module_perso',
+    validateFiltresActionsPilotes: 'tdb_valider_filtres_actions_pilotes',
+    validateFiltresActionsModifiees: 'tdb_valider_filtres_actions_modifiees',
+  },
+  recherches: {
+    updateFiltresTypePlan: 'collectivites_onglet_pa:filtre_type_pa_select',
+    updateFiltresLabellisation:
+      'collectivites_onglet_referentiels:filtre_labellisation_select',
+    viewContacts: 'collectivites:voir_contacts_click',
+    viewCollectivite: 'collectivites_onglet_collectivites:cartes_click',
+    viewPlan: 'collectivites_onglet_pa:cartes_click',
+    viewReferentiel: 'collectivites_onglet_referentiels:cartes_click',
+  },
+  referentiels: {
+    exportScore: 'export_score',
+    exportAuditScore: 'export_audit_score',
+    viewIndicateurs: 'view_indicateurs_click',
+    viewLabellisation: 'cta_labellisation',
+    startEtatLieux: 'cta_edl_commencer',
+    personalizeReferentiel: 'cta_edl_personnaliser',
+    openSidePanel: 'side_panel_open',
+    clickSommaireInfosPanel: 'sommaire_infos_panel_click',
+  },
+  viewGraphZoom: 'zoom_graph',
+  downloadGraph: 'export_graph',
+  togglePanel: 'panneau_lateral_toggle',
+  toggleNavigationLaterale: 'navigation_laterale_toggle',
+  accueil: {
+    viewSyntheseEtatLieux: 'accueil:edl_synthese_click',
+    viewMoreInfos: 'accueil:edl_plus_infos_click',
+    viewTableauDeBordPersonnel: 'accueil:PA_mon_suivi_personnel_click',
+    viewTableauDeBordCollectivite:
+      'accueil:PA_tableau_de_bord_collectivite_click',
+    createPlan: 'accueil:PA_creer_plan_click',
+    createFiche: 'accueil:PA_creer_fiche_click',
+    viewToutesLesFiches: 'accueil:PA_toutes_les_fiches_click',
+    viewTousLesIndicateurs: 'accueil:indicateurs_tous_click',
+    viewIndicateursOpenData: 'accueil:indicateurs_open_data_click',
+    viewTrajectoires: 'accueil:trajectoires_decouvrir_click',
+    viewCollectivites: 'accueil:collectivites_inspiration_click',
+    viewCollectivitesConfidentialite:
+      'accueil:collectivites_confidentialite_click',
+    viewPanierActions: 'accueil:panier_actions_tester_click',
+    viewSite: 'accueil:retourner_site_click',
+  },
+} as const;
+
+// Get all values of nested objects (i.e filtres_update)
+export type EventName = {
+  [K in keyof typeof Event]: (typeof Event)[K] extends Record<string, unknown>
+    ? (typeof Event)[K][keyof (typeof Event)[K]]
+    : (typeof Event)[K];
+}[keyof typeof Event];
